pyproject.toml
[project]
name = "agent-hack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.16",
    "ruff>=0.11.12",
]

